+++ app.py
@@ -155,20 +155,20 @@ with gr.Blocks(title=init.ver, theme=gr.themes.Soft(primary_hue="red", secondary
     
     clear_all_button.click(init.clear_all, outputs=clear_all_tb)
 
-if init.debug is True:
+if config.debug is True:
     init.delete_tmp_pngs()
-    if init.preload_models is True:
+    if config.preload_models is True:
         init.preloader()
     end_time = time.time()
     total_time = round(end_time - start_time)
     clear()
     print(f"Executing init time: {total_time}s")
     multiai.queue()
-    multiai.launch(inbrowser=init.inbrowser, share=init.share_gradio)
-elif init.debug is False:
+    multiai.launch(inbrowser=config.inbrowser, share=config.share_gradio)
+elif config.debug is False:
     init.delete_tmp_pngs()
-    if init.preload_models is True:
+    if config.preload_models is True:
         init.preloader() 
     clear()
     multiai.queue()
-    multiai.launch(inbrowser=init.inbrowser, share=init.share_gradio)
+    multiai.launch(inbrowser=config.inbrowser, share=config.share_gradio)